Update example to use one page
