--- src/assets/css/a11y/a11y.css
@@ -1,92 +1,0 @@
-html {
-	padding: 0;
-	margin: 0;
-}
-
-body {
-	padding: 0;
-	margin: 0;
-}
-
-.sronly {
-	position: absolute;
-	width: 1px;
-	height: 1px;
-	padding: 0;
-	margin: -1px;
-	overflow: hidden;
-	clip: rect( 0, 0, 0, 0 );
-	border: 0;
-}
-
-.bar {
-	border-bottom: 4px solid #f5f5f5;
-	border-bottom: .25rem solid #f5f5f5;
-}
-
-.a11y-button {
-	display: inline-block;
-	-webkit-appearance: none;
-	-moz-appearance: none;
-	appearance: none;
-	font-family: -apple-system,BlinkMacSystemFont,"Segoe UI","Roboto","Oxygen","Ubuntu","Cantarell","Fira Sans","Droid Sans","Helvetica Neue",sans-serif;
-	color: #007099;
-	background-color: #fff;
-	cursor: pointer;
-	vertical-align: middle;
-	font-size: 14px;
-	font-size: .875rem;
-	line-height: 1.14286;
-	padding: 12px 16px 12px;
-	padding: .75rem 1rem .75rem;
-	border: none;
-	border-bottom: 4px solid #f5f5f5;
-	border-bottom: .25rem solid #f5f5f5;
-	margin-bottom: -4px;
-	margin-bottom: -0.25rem;
-}
-
-.a11y-button:focus {
-	outline: 3px solid #9263DE;
-	outline-offset: 2px;
-}
-
-.a11y-button.is-active {
-	border-color: #007099;
-}
-
-.a11y-button.a11y-button--soft {
-	color: #313131;
-}
-
-.a11y .example .bar,
-.a11y .example .content,
-.a11y .example .bar,
-.a11y .example .content {
-	display: none;
-}
-
-@supports ( filter: url( "filters.svg#deuteranopia" ) ){
-	.filter-fail {
-		display: none;
-	}
-
-	.a11y .example .bar,
-	.a11y .example .content,
-	.a11y .example .bar,
-	.a11y .example .content {
-		display: block;
-	}
-}
-
-
-.svg-filter {
-	position: absolute;
-	width: 1px;
-	height: 1px;
-	padding: 0;
-	margin: -1px;
-	overflow: hidden;
-	clip: rect(0,0,0,0);
-	border: 0;
-}
