--- src/assets/css/a11y/deuteranopia.css
@@ -1,1 +1,0 @@
-.filter{filter:url("#deuteranopia")}
