--- src/assets/css/a11y/grayscale.css
@@ -1,1 +1,0 @@
-.filter{filter:grayscale(100%)}
