--- src/assets/css/a11y/tritanopia.css
@@ -1,1 +1,0 @@
-.filter{filter:url("#tritanopia")}
