+++ src/js/a11y/030-filter.js
@@ -1,15 +1,12 @@
 var filters = document.querySelector( '.js-filter' );
 if( filters !== null ) {
-	var active = document.getElementById( 'filter-stylesheet' ).getAttribute( 'href' ).split( '/' );
-	active = active[ active.length - 1 ].replace( '.css', '' );
-
 	filters.insertAdjacentHTML(
 		'afterbegin',
 		'<div class="bar">' +
-		'<button type="button" class="a11y-button js-filter-btn' + ( active == 'deuteranopia' ? ' is-active' : '' ) + '" data-filter="deuteranopia">' +
+		'<button type="button" class="a11y-button js-filter-btn is-active" data-filter="deuteranopia">' +
 			'<span class="sronly">Apply </span>Deuteranopia<span class="sronly"> color filter</span>' +
 		'</button>' +
-		'<button type="button" class="a11y-button js-filter-btn' + ( active == 'tritanopia' ? ' is-active' : '' ) + '" data-filter="tritanopia">' +
+		'<button type="button" class="a11y-button js-filter-btn" data-filter="tritanopia">' +
 			'<span class="sronly">Apply </span>Tritanopia<span class="sronly"> color filter</span>' +
 		'</button>' +
 		'</div>'
@@ -17,16 +14,15 @@ if( filters !== null ) {
 }
 
 AddEvent( document.querySelectorAll( '.js-filter-btn' ), 'click', function( event, $this ) {
-	var sheet = document.getElementById( 'filter-stylesheet' );
-	var filter = $this.getAttribute( 'data-filter' );
-	var url = sheet.getAttribute( 'data-' + filter );
-
-	sheet.setAttribute( 'href', url );
+	var newFilter = 'js-filter--' + $this.getAttribute( 'data-filter' );
 
 	var allButtons = document.querySelectorAll( '.js-filter-btn' );
 	for( var i = 0; i < allButtons.length; i++ ) {
+		var oldFilter = 'js-filter--' + allButtons[ i ].getAttribute( 'data-filter' );
+		RemoveClass( filters, oldFilter );
 		RemoveClass( allButtons[ i ], 'is-active' );
 	}
 
+	AddClass( filters, newFilter );
 	AddClass( $this, 'is-active' );
 });
