+++ src/layout/download/furnace.js
@@ -1,6 +1,6 @@
-import { AUcheckbox, AUradio } from '../../_uikit/layout/control-input';
+import { AUradio } from '../../_uikit/layout/control-input';
 import GetData from './../../helper/getData';
-import Code from './../code';
+import CodeSnippet from './../code-snippet';
 
 import PropTypes from 'prop-types';
 import React, { Fragment } from 'react';
@@ -140,7 +140,7 @@ const Furnace = ({ components, _ID, _body, _parseYaml, _relativeURL }) => {
 
 					<div className="furnace-npm">
 						<h4>npm</h4>
-						<Code aria-live="polite" className="js-furnace-code">npm install --save @gov.au/core</Code>
+						<CodeSnippet aria-live="polite" className="js-furnace-code">npm install --save @gov.au/core</CodeSnippet>
 					</div>
 
 				</div>
