+++ src/layout/example/page.js
@@ -1,7 +1,7 @@
-import PropTypes from 'prop-types';
 import React from 'react';
-import Path from 'path';
-import Fs from 'fs';
+
+import GetModule             from '../../helper/getModule';
+import { GetComponentValue } from '../../helper/getData';
 
 
 /**
@@ -9,7 +9,26 @@ import Fs from 'fs';
  *
  * @disable-docs
  */
-const Example = ({ _ID, _relativeURL, pagetitle, example, alignment, fullwidth }) => {
+const Example = ({
+	pagetitle,
+	filter,
+	tabbing,
+	examples,
+	fullwidth,
+	page,
+	alignment,
+	_ID,
+	_relativeURL,
+	_parseYaml,
+	_parents,
+	_pages
+}) => {
+	const componentID = _ID.split( '/' )[ 1 ];
+
+	const a11yPage = tabbing || filter ? true : false;
+
+	const module = GetModule( _parents, _pages, _ID );
+	const componentTitle = GetComponentValue( module, 'name', _parseYaml );
 
 	const headContent = `
 <meta charset="utf-8">
@@ -33,13 +52,13 @@ const Example = ({ _ID, _relativeURL, pagetitle, example, alignment, fullwidth }
 <meta name="twitter:name" content="Australian Government Design System">
 <meta name="twitter:image" content="https://designsystem.gov.au/assets/favicons/designsystem.jpg">
 <meta property="og:type" content="website">
-<meta property="og:title" content="${ pagetitle } - Australian Government Design System">
+<meta property="og:title" content="${ pagetitle ? pagetitle : componentTitle + ` example` } - Australian Government Design System">
 <meta property="og:site_name" content="Australian Government Design System">
 <meta property="og:description" content="Inclusive design, open-source code and shared insights">
 <meta property="og:image" content="https://designsystem.gov.au/assets/favicons/designsystem.jpg">
 <meta property="og:url" content="https://designsystem.gov.au">
 
-<title>${ pagetitle } - Australian Government Design System</title>
+<title>${ pagetitle ? pagetitle : componentTitle + ` example` } - Australian Government Design System</title>
 
 <link rel="stylesheet" href=${ _relativeURL( '/assets/css/example.css', _ID ) } />
 
@@ -47,32 +66,157 @@ const Example = ({ _ID, _relativeURL, pagetitle, example, alignment, fullwidth }
 	<script src="${ _relativeURL( '/assets/js/html5shiv.js', _ID ) }"></script>
 	<script src="${ _relativeURL( '/assets/js/respond.js', _ID ) }"></script>
 <![endif]-->
-
 <script>var $html=document.documentElement;if($html.classList)$html.classList.remove("no-js"),$html.classList.add("js");else{var className="no-js";$html.className=$html.className.replace(new RegExp("(^|\\b)"+className.split(" ").join("|")+"(\\b|$)","gi")," "),$html.className+=" js"}</script>`;
+
+	let exampleLayout = Object.values( examples ).map( ( example, key ) => {
+		if( example.style ){
+			return (
+				<div className={ `example example-${ componentID } example--${ example.style }` } key={ key }>
+					{ example.template }
+				</div>
+			);
+		}
+
+		return (
+			<div className={ `example  example-${ componentID }` } key={ key }>
+				{ example.template }
+			</div>
+		);
+	});
+
+	// Change the alignment if inline
+	if( alignment === 'inline' && exampleLayout.length == 2 ){
+		exampleLayout =
+			<div className="row">
+				{ exampleLayout.map( ( example, key ) => (
+					<div className="col-sm-6" key={ key }>{ example }</div>
+				) ) }
+			</div>
+	}
+
+	const a11yFilter = <svg
+		aria-hidden="true"
+		className="svg-filter"
+		xmlns="http://www.w3.org/2000/svg"
+		version="1.1">
+		<defs>
+			<filter id="protanopia">
+				<feColorMatrix
+					in="SourceGraphic"
+					type="matrix"
+					values="0.567, 0.433, 0,     0, 0
+									0.558, 0.442, 0,     0, 0
+									0,     0.242, 0.758, 0, 0
+									0,     0,     0,     1, 0"/>
+			</filter>
+			<filter id="protanomaly">
+				<feColorMatrix
+					in="SourceGraphic"
+					type="matrix"
+					values="0.817, 0.183, 0,     0, 0
+									0.333, 0.667, 0,     0, 0
+									0,     0.125, 0.875, 0, 0
+									0,     0,     0,     1, 0"/>
+			</filter>
+			<filter id="deuteranopia">
+				<feColorMatrix
+					in="SourceGraphic"
+					type="matrix"
+					values="0.625, 0.375, 0,   0, 0
+									0.7,   0.3,   0,   0, 0
+									0,     0.3,   0.7, 0, 0
+									0,     0,     0,   1, 0"/>
+			</filter>
+			<filter id="deuteranomaly">
+				<feColorMatrix
+					in="SourceGraphic"
+					type="matrix"
+					values="0.8,   0.2,   0,     0, 0
+									0.258, 0.742, 0,     0, 0
+									0,     0.142, 0.858, 0, 0
+									0,     0,     0,     1, 0"/>
+			</filter>
+			<filter id="tritanopia">
+				<feColorMatrix
+					in="SourceGraphic"
+					type="matrix"
+					values="0.95, 0.05,  0,     0, 0
+									0,    0.433, 0.567, 0, 0
+									0,    0.475, 0.525, 0, 0
+									0,    0,     0,     1, 0"/>
+			</filter>
+			<filter id="tritanomaly">
+				<feColorMatrix
+					in="SourceGraphic"
+					type="matrix"
+					values="0.967, 0.033, 0,     0, 0
+									0,     0.733, 0.267, 0, 0
+									0,     0.183, 0.817, 0, 0
+									0,     0,     0,     1, 0"/>
+			</filter>
+			<filter id="achromatopsia">
+				<feColorMatrix
+					in="SourceGraphic"
+					type="matrix"
+					values="0.299, 0.587, 0.114, 0, 0
+									0.299, 0.587, 0.114, 0, 0
+									0.299, 0.587, 0.114, 0, 0
+									0,     0,     0,     1, 0"/>
+			</filter>
+			<filter id="achromatomaly">
+				<feColorMatrix
+					in="SourceGraphic"
+					type="matrix"
+					values="0.618, 0.320, 0.062, 0, 0
+									0.163, 0.775, 0.062, 0, 0
+									0.163, 0.320, 0.516, 0, 0
+									0,     0,     0,     1, 0"/>
+			</filter>
+		</defs>
+	</svg>;
+
 
 	return (
 		<html>
-		<head dangerouslySetInnerHTML={{ __html: headContent }} />
-		<body className="au-body au-grid">
-			<main className={ `example${ alignment === 'center' ? ' example--center' : '' }${ fullwidth === true ? ' example--fullwidth' : '' }` }>
-				<div className='content'>
-					{ example }
+		<head dangerouslySetInnerHTML={{ __html: headContent }}  />
+		<body className={ `au-grid${ a11yPage ? ' a11y' : '' }${ page ? ' live-demo' : '' }` }>
+			{
+				page
+					? <header className="header__example">
+							<div className="container-fluid">
+								<div className="row">
+									<div className="col-md-12">
+										<a className="au-direction-link au-direction-link--left" href={ _relativeURL( `/components/${ module }`, _ID ) }>
+											View { componentTitle } overview
+										</a>
+									</div>
+								</div>
+							</div>
+						</header>
+					: ''
+			}
+			{ filter ? a11yFilter : '' }
+			<main className={`${ fullwidth === false ? ' example--align-middle' : '' }${ tabbing ? ' js-tabbing' : '' }${ filter === 'protanopia' || filter === 'deuteranopia' ? ' js-filter' : '' }${ filter ? ` js-filter--${ filter}` : '' }`}>
+				<div className={`content${ tabbing ? ' js-tabbing-area' : '' }`}>
+					{ exampleLayout }
 				</div>
+				{
+					filter
+						? <p className="filter-fail">Unfortunately this feature is not available in this browser.</p>
+						: ''
+				}
 			</main>
-
-			<script src={ _relativeURL( '/assets/js/footer.js', _ID ) } />
 			<script src={ _relativeURL( '/assets/js/iframe-resizer-contentWindow.js', _ID ) } />
+			<script src={ _relativeURL( '/assets/js/footer.js', _ID ) } />
+			{ a11yPage ? <script src={ _relativeURL( '/assets/js/a11y.js', _ID ) } /> : '' }
 		</body>
 		</html>
-	)
+		)
 };
 
 Example.propTypes = {
 };
 
-Example.defaultProps = {
-	alignment: 'left',
-	fullwidth: false
-};
+Example.defaultProps = {};
 
 export default Example;
