--- src/layout/live/page.js
@@ -1,94 +1,0 @@
-import GetModule             from '../../helper/getModule';
-import { GetComponentValue } from '../../helper/getData';
-
-import PropTypes from 'prop-types';
-import React from 'react';
-import Path from 'path';
-import Fs from 'fs';
-
-
-/**
- * The example component
- *
- * @disable-docs
- */
-const Live = ({ _ID, _parents, _pages, _relativeURL, _parseYaml, pagetitle, example, alignment, fullwidth }) => {
-	const headContent = `
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width">
-<meta http-equiv="x-ua-compatible" content="ie=edge">
-
-<link rel="apple-touch-icon" sizes="180x180" href=${ _relativeURL('/assets/favicons/apple-touch-icon.png', _ID ) }>
-<link rel="icon" type="image/png" sizes="32x32" href=${ _relativeURL('/assets/favicons/favicon-32x32.png', _ID ) }>
-<link rel="icon" type="image/png" sizes="16x16" href=${ _relativeURL('/assets/favicons/favicon-16x16.png', _ID ) }>
-<link rel="manifest" href=${ _relativeURL('/assets/favicons/site.webmanifest', _ID ) }>
-<link rel="mask-icon" href=${ _relativeURL('/assets/favicons/safari-pinned-tab.svg', _ID ) } color="#06262d">
-<link rel="shortcut icon" href=${ _relativeURL('/assets/favicons/favicon.ico', _ID ) }>
-<meta name="msapplication-config" content=${ _relativeURL('/assets/favicons/browserconfig.xml', _ID ) }>
-<meta name="msapplication-TileColor" content="#ffffff">
-<meta name="theme-color" content="#ffffff">
-<meta name="robots" content="index, follow">
-<meta name="author" content="Digital Transformation Agency">
-<meta name="twitter:card" content="summary_large_image">
-<meta name="twitter:title" content="Australian Government Design System">
-<meta name="twitter:description" content="Inclusive design, open-source code and shared insights">
-<meta name="twitter:name" content="Australian Government Design System">
-<meta name="twitter:image" content="https://designsystem.gov.au/assets/favicons/designsystem.jpg">
-<meta property="og:type" content="website">
-<meta property="og:title" content="${ pagetitle } - Australian Government Design System">
-<meta property="og:site_name" content="Australian Government Design System">
-<meta property="og:description" content="Inclusive design, open-source code and shared insights">
-<meta property="og:image" content="https://designsystem.gov.au/assets/favicons/designsystem.jpg">
-<meta property="og:url" content="https://designsystem.gov.au">
-
-<title>${ pagetitle } - Australian Government Design System</title>
-
-<link rel="stylesheet" href=${ _relativeURL( '/assets/css/example.css', _ID ) } />
-
-<!--[if lte IE 9]>
-	<script src="${ _relativeURL( '/assets/js/html5shiv.js', _ID ) }"></script>
-	<script src="${ _relativeURL( '/assets/js/respond.js', _ID ) }"></script>
-<![endif]-->
-
-<script src=${ _relativeURL( '/assets/js/header.js', _ID ) }></script>`;
-
-	const module = GetModule( _parents, _pages, _ID );
-
-	return (
-		<html>
-		<head dangerouslySetInnerHTML={{ __html: headContent }} />
-		<body className={`live-demo au-body au-grid`} >
-			<noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-5Z7S8GB" height="0" width="0" style={{display:'none',visibility:'hidden'}}></iframe></noscript>
-			<header className="header__example">
-				<div className="container-fluid">
-					<div className="row">
-						<div className="col-md-12">
-							<a className="au-direction-link au-direction-link--left" href={ _relativeURL( `/components/${ module }`, _ID ) }>
-								View { GetComponentValue( module, 'name', _parseYaml ) } overview
-							</a>
-						</div>
-					</div>
-				</div>
-			</header>
-
-			<div className={ `example${ alignment === 'center' ? ' example--center' : '' }${ fullwidth ? ' example--fullwidth' : '' }` }>
-				<div className='content'>
-					{ example }
-				</div>
-			</div>
-
-			<script src={ _relativeURL( '/assets/js/footer.js', _ID ) } />
-			<script src={ _relativeURL( '/assets/js/iframe-resizer-contentWindow.js', _ID ) } />
-		</body>
-		</html>
-	)
-};
-
-Live.propTypes = {
-};
-
-Live.defaultProps = {
-	alignment: 'left'
-};
-
-export default Live;
