--- src/sass/hotfix.scss
@@ -1,4 +1,0 @@
-// This file contains hotfixes we want to replace
-[tabindex="-1"] {
-	outline: none;
-}
