+++ src/sass/style.scss
@@ -51,7 +51,6 @@
 @import 'components/callout-with-button';
 @import 'components/section';
 @import 'components/searchpage';
-@import 'hotfix';
 
 //Possible Visual Design Tweaks to UIKit
 @import 'components/vd-tweaks';
